+++ src/file_drop.rs
@@ -1,5 +1,0 @@
-// Copyright 2020-2023 Tauri Programme within The Commons Conservancy
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-License-Identifier: MIT
-
-use std::path::PathBuf;