+++ src/webview/android/mod.rs
@@ -29,6 +29,12 @@ pub(crate) mod binding;
 mod main_pipe;
 use main_pipe::{CreateWebViewAttributes, MainPipe, WebViewMessage, MAIN_PIPE};
 
+pub struct Context<'a> {
+  pub env: JNIEnv<'a>,
+  pub activity: JObject<'a>,
+  pub webview: JObject<'a>,
+}
+
 #[macro_export]
 macro_rules! android_binding {
   ($domain:ident, $package:ident, $main: ident) => {
@@ -51,6 +57,22 @@ macro_rules! android_binding {
       [JObject],
       jobject
     );
+    android_fn!(
+      $domain,
+      $package,
+      RustWebViewClient,
+      withAssetLoader,
+      [],
+      jboolean
+    );
+    android_fn!(
+      $domain,
+      $package,
+      RustWebViewClient,
+      assetLoaderDomain,
+      [],
+      jstring
+    );
     android_fn!($domain, $package, Ipc, ipc, [JString]);
     android_fn!(
       $domain,
@@ -65,6 +87,8 @@ macro_rules! android_binding {
 pub static IPC: OnceCell<UnsafeIpc> = OnceCell::new();
 pub static REQUEST_HANDLER: OnceCell<UnsafeRequestHandler> = OnceCell::new();
 pub static TITLE_CHANGE_HANDLER: OnceCell<UnsafeTitleHandler> = OnceCell::new();
+pub static WITH_ASSET_LOADER: OnceCell<bool> = OnceCell::new();
+pub static ASSET_LOADER_DOMAIN: OnceCell<String> = OnceCell::new();
 
 pub struct UnsafeIpc(Box<dyn Fn(&Window, String)>, Rc<Window>);
 impl UnsafeIpc {
@@ -98,7 +122,7 @@ unsafe impl Sync for UnsafeTitleHandler {}
 pub unsafe fn setup(env: JNIEnv, looper: &ForeignLooper, activity: GlobalRef) {
   // we must create the WebChromeClient here because it calls `registerForActivityResult`,
   // which gives an `LifecycleOwners must call register before they are STARTED.` error when called outside the onCreate hook
-  let rust_webchrome_client_class = find_my_class(
+  let rust_webchrome_client_class = find_class(
     env,
     activity.as_obj(),
     format!("{}/RustWebChromeClient", PACKAGE.get().unwrap()),
@@ -144,7 +168,7 @@ impl InnerWebView {
   pub fn new(
     window: Rc<Window>,
     attributes: WebViewAttributes,
-    _pl_attrs: super::PlatformSpecificWebViewAttributes,
+    pl_attrs: super::PlatformSpecificWebViewAttributes,
     _web_context: Option<&mut WebContext>,
   ) -> Result<Self> {
     let WebViewAttributes {
@@ -158,6 +182,12 @@ impl InnerWebView {
       headers,
       ..
     } = attributes;
+
+    let super::PlatformSpecificWebViewAttributes {
+      on_webview_created,
+      with_asset_loader,
+      asset_loader_domain,
+    } = pl_attrs;
 
     if let Some(u) = url {
       let mut url_string = String::from(u.as_str());
@@ -175,7 +205,13 @@ impl InnerWebView {
         background_color,
         transparent,
         headers,
+        on_webview_created,
       }));
+    }
+
+    WITH_ASSET_LOADER.get_or_init(move || with_asset_loader);
+    if let Some(domain) = asset_loader_domain {
+      ASSET_LOADER_DOMAIN.get_or_init(move || domain);
     }
 
     REQUEST_HANDLER.get_or_init(move || {
@@ -197,8 +233,18 @@ impl InnerWebView {
             .unwrap();
 
           if let Ok(mut response) = (custom_protocol.1)(&request) {
-            if response.headers().get(CONTENT_TYPE) == Some(&HeaderValue::from_static("text/html"))
-            {
+            let should_inject_scripts = response
+              .headers()
+              .get(CONTENT_TYPE)
+              // Content-Type must begin with the media type, but is case-insensitive.
+              // It may also be followed by any number of semicolon-delimited key value pairs.
+              // We don't care about these here.
+              // source: https://httpwg.org/specs/rfc9110.html#rfc.section.8.3.1
+              .and_then(|content_type| content_type.to_str().ok())
+              .map(|content_type_str| content_type_str.to_lowercase().starts_with("text/html"))
+              .unwrap_or_default();
+
+            if should_inject_scripts {
               if !initialization_scripts.is_empty() {
                 let mut document =
                   kuchiki::parse_html().one(String::from_utf8_lossy(response.body()).into_owned());
@@ -332,7 +378,8 @@ fn hash_script(script: &str) -> String {
   format!("'sha256-{}'", base64::encode(hash))
 }
 
-fn find_my_class<'a>(
+/// Finds a class in the project scope.
+pub fn find_class<'a>(
   env: JNIEnv<'a>,
   activity: JObject<'a>,
   name: String,
@@ -347,6 +394,16 @@ fn find_my_class<'a>(
     )?
     .l()?;
   Ok(my_class.into())
+}
+
+/// Dispatch a closure to run on the Android context.
+///
+/// The closure takes the JNI env, the Android activity instance and the possibly null webview.
+pub fn dispatch<F>(func: F)
+where
+  F: FnOnce(JNIEnv, JObject, JObject) + Send + 'static,
+{
+  MainPipe::send(WebViewMessage::Jni(Box::new(func)));
 }
 
 fn create_headers_map<'a, 'b>(