+++ src/webview/webkitgtk/mod.rs
@@ -267,10 +267,17 @@ impl InnerWebView {
     }
     webview.grab_focus();
 
+    if let Some(context) = WebViewExt::context(&*webview) {
+      use webkit2gtk::WebContextExt;
+      context.set_use_system_appearance_for_scrollbars(false);
+    }
+
     // Enable webgl, webaudio, canvas features as default.
     if let Some(settings) = WebViewExt::settings(&*webview) {
       settings.set_enable_webgl(true);
       settings.set_enable_webaudio(true);
+      settings
+        .set_enable_back_forward_navigation_gestures(attributes.back_forward_navigation_gestures);
 
       // Enable clipboard
       if attributes.clipboard {