+++ src/webview/webview2/file_drop.rs
@@ -19,14 +19,15 @@ use std::{
 };
 
 use windows::Win32::{
-  Foundation::{self as win32f, BOOL, DRAGDROP_E_INVALIDHWND, HWND, LPARAM, POINTL},
+  Foundation::{self as win32f, BOOL, DRAGDROP_E_INVALIDHWND, HWND, LPARAM, POINT, POINTL},
+  Graphics::Gdi::ScreenToClient,
   System::{
     Com::{IDataObject, DVASPECT_CONTENT, FORMATETC, TYMED_HGLOBAL},
     Ole::{
-      IDropTarget, IDropTarget_Impl, RegisterDragDrop, RevokeDragDrop, DROPEFFECT_COPY,
-      DROPEFFECT_NONE,
+      IDropTarget, IDropTarget_Impl, RegisterDragDrop, RevokeDragDrop, CF_HDROP, DROPEFFECT,
+      DROPEFFECT_COPY, DROPEFFECT_NONE,
     },
-    SystemServices::CF_HDROP,
+    SystemServices::MODIFIERKEYS_FLAGS,
   },
   UI::{
     Shell::{DragFinish, DragQueryFileW, HDROP},
@@ -36,7 +37,9 @@ use windows::Win32::{
 
 use windows_implement::implement;
 
-use crate::application::window::Window;
+use crate::application::{
+  dpi::PhysicalPosition, platform::windows::WindowExtWindows, window::Window,
+};
 
 pub(crate) struct FileDropController {
   drop_targets: Vec<IDropTarget>,
@@ -107,7 +110,7 @@ unsafe extern "system" fn enumerate_callback(hwnd: HWND, lparam: LPARAM) -> BOOL
 pub struct FileDropHandler {
   window: Rc<Window>,
   listener: Rc<dyn Fn(&Window, FileDropEvent) -> bool>,
-  cursor_effect: UnsafeCell<u32>,
+  cursor_effect: UnsafeCell<DROPEFFECT>,
   hovered_is_valid: UnsafeCell<bool>, /* If the currently hovered item is not valid there must not be any `HoveredFileCancelled` emitted */
 }
 
@@ -119,7 +122,7 @@ impl FileDropHandler {
     Self {
       window,
       listener,
-      cursor_effect: DROPEFFECT_NONE.0.into(),
+      cursor_effect: DROPEFFECT_NONE.into(),
       hovered_is_valid: false.into(),
     }
   }
@@ -145,13 +148,13 @@ impl FileDropHandler {
         let hdrop = HDROP(medium.Anonymous.hGlobal);
 
         // The second parameter (0xFFFFFFFF) instructs the function to return the item count
-        let item_count = DragQueryFileW(hdrop, 0xFFFFFFFF, &mut []);
+        let item_count = DragQueryFileW(hdrop, 0xFFFFFFFF, None);
 
         for i in 0..item_count {
           // Get the length of the path string NOT including the terminating null character.
           // Previously, this was using a fixed size array of MAX_PATH length, but the
           // Windows API allows longer paths under certain circumstances.
-          let character_count = DragQueryFileW(hdrop, i, &mut []) as usize;
+          let character_count = DragQueryFileW(hdrop, i, None) as usize;
           let str_len = character_count + 1;
 
           // Fill path_buf with the null-terminated file name
@@ -187,9 +190,9 @@ impl IDropTarget_Impl for FileDropHandler {
   fn DragEnter(
     &self,
     pDataObj: &Option<IDataObject>,
-    _grfKeyState: u32,
-    _pt: &POINTL,
-    pdwEffect: *mut u32,
+    _grfKeyState: MODIFIERKEYS_FLAGS,
+    pt: &POINTL,
+    pdwEffect: *mut DROPEFFECT,
   ) -> windows::core::Result<()> {
     let mut paths = Vec::new();
     unsafe {
@@ -200,21 +203,30 @@ impl IDropTarget_Impl for FileDropHandler {
       } else {
         DROPEFFECT_NONE
       };
-      *pdwEffect = cursor_effect.0;
+      *pdwEffect = cursor_effect;
       *self.hovered_is_valid.get() = hovered_is_valid;
-      *self.cursor_effect.get() = cursor_effect.0;
+      *self.cursor_effect.get() = cursor_effect;
+
+      let mut pt = POINT { x: pt.x, y: pt.y };
+      ScreenToClient(HWND(self.window.hwnd() as _), &mut pt);
     }
 
-    (self.listener)(&self.window, FileDropEvent::Hovered(paths));
+    (self.listener)(
+      &self.window,
+      FileDropEvent::Hovered {
+        paths,
+        position: PhysicalPosition::new(pt.x as _, pt.y as _),
+      },
+    );
 
     Ok(())
   }
 
   fn DragOver(
     &self,
-    _grfKeyState: u32,
+    _grfKeyState: MODIFIERKEYS_FLAGS,
     _pt: &POINTL,
-    pdwEffect: *mut u32,
+    pdwEffect: *mut DROPEFFECT,
   ) -> windows::core::Result<()> {
     unsafe { *pdwEffect = *self.cursor_effect.get() };
     Ok(())
@@ -230,9 +242,9 @@ impl IDropTarget_Impl for FileDropHandler {
   fn Drop(
     &self,
     pDataObj: &Option<IDataObject>,
-    _grfKeyState: u32,
-    _pt: &POINTL,
-    _pdwEffect: *mut u32,
+    _grfKeyState: MODIFIERKEYS_FLAGS,
+    pt: &POINTL,
+    _pdwEffect: *mut DROPEFFECT,
   ) -> windows::core::Result<()> {
     let mut paths = Vec::new();
     unsafe {
@@ -240,9 +252,18 @@ impl IDropTarget_Impl for FileDropHandler {
       if let Some(hdrop) = hdrop {
         DragFinish(hdrop);
       }
+
+      let mut pt = POINT { x: pt.x, y: pt.y };
+      ScreenToClient(HWND(self.window.hwnd() as _), &mut pt);
     }
 
-    (self.listener)(&self.window, FileDropEvent::Dropped(paths));
+    (self.listener)(
+      &self.window,
+      FileDropEvent::Dropped {
+        paths,
+        position: PhysicalPosition::new(pt.x as _, pt.y as _),
+      },
+    );
 
     Ok(())
   }