+++ src/webview/wkwebview/mod.rs
@@ -347,6 +347,10 @@ impl InnerWebView {
         // set all autoresizingmasks
         let () = msg_send![webview, setAutoresizingMask: 31];
         let _: () = msg_send![webview, initWithFrame:frame configuration:config];
+
+        // disable scroll bounce by default
+        let scroll: id = msg_send![webview, scrollView];
+        let _: () = msg_send![scroll, setBounces: NO];
       }
 
       // allowsBackForwardNavigation
@@ -614,7 +618,7 @@ impl InnerWebView {
               );
               cls.register()
             }
-            None => class!(UIViewController),
+            None => class!(WryDownloadDelegate),
           };
 
           let download_delegate: id = msg_send![cls, new];